requirements.txt
banal==1.0.6
normality==2.5.0
pantomime==0.6.1
./followthemoney
followthemoney-store[postgresql]==3.0.6
servicelayer[google,amazon]==1.22.1
languagecodes==1.1.1
countrytagger==0.1.2
pyicu==2.12
google-cloud-vision==3.5.0
tesserocr==2.6.2
spacy==3.6.1
fingerprints==1.1.1
fasttext==0.9.2

# Development
pytest==7.4.4
pytest-cov==4.1.0
click==8.1.6

# File format support
dbf==0.99.9
pymediainfo==6.1.0
python-magic==0.4.27
rarfile==4.1
xlrd==2.0.1
openpyxl==3.1.2
odfpy==1.4.1
cchardet==2.1.7
lxml==5.0.0
olefile==0.47
Pillow==10.1.0
vobject==0.9.6.1
msglite==0.30.0
icalendar==5.0.11

cryptography==41.0.7
requests[security]==2.31.0
pymupdf==1.21.1

sentry_sdk==1.39.1
